load-tests/requirements.txt
locust>=2.29.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
